[Style] Modifying the overall
