+++ public/index.html
@@ -87,6 +87,20 @@
       -moz-osx-font-smoothing: grayscale;
       -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
       background-color: #000312;
+
+      ::-webkit-scrollbar {
+        background-color: transparent;
+      }
+      
+      ::-webkit-scrollbar-thumb {
+        border: 1px solid #98A1C03d;
+        background-color: #131a2a;
+      }
+      
+      ::-webkit-scrollbar-track {
+        border-left: 1px solid #98A1C03d;
+        background-color: transparent;
+      }
     }
 
     /* Use this to apply network-specific gradient backgrounds, in RadialGradientByChainUpdater.ts */
@@ -114,11 +128,11 @@
       }
     } */
 
-    @media (prefers-color-scheme: light) {
+    /* @media (prefers-color-scheme: light) {
       html {
         background: radial-gradient(100% 100% at 50% 0%, rgba(255, 184, 226, 0.51) 0%, rgba(255, 255, 255, 0) 100%), #FFFFFF
       }
-    }
+    } */
   </style>
 </head>
