+++ src/components/PositionTable/LeveragePositionTable/TokenTable.tsx
@@ -9,6 +9,7 @@ import { useLocation } from 'react-router-dom'
 import styled from 'styled-components/macro'
 import { LimitlessPositionDetails } from 'types/leveragePosition'
 
+import { TokenDataContainer } from '../comonStyle'
 import { MAX_WIDTH_MEDIA_BREAKPOINT } from './constants'
 import { filterStringAtom, PositionSortMethod, sortAscendingAtom, sortMethodAtom } from './state'
 import { HeaderRow, LoadedRow, LoadingRow } from './TokenRow'
@@ -16,26 +17,11 @@ import { HeaderRow, LoadedRow, LoadingRow } from './TokenRow'
 const GridContainer = styled.div`
   display: flex;
   flex-direction: column;
-  // margin-top: 8px;
-
   width: 100%;
   /* max-width: ${MAX_WIDTH_MEDIA_BREAKPOINT}; */
   background-color: ${({ theme }) => theme.backgroundSurface};
-  box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.01), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04),
-    0px 24px 32px rgba(0, 0, 0, 0.01);
-  /* margin-left: auto;
-  margin-right: auto; */
   justify-content: flex-start;
   align-items: flex-start;
-  // border: 1px solid ${({ theme }) => theme.backgroundOutline};
-`
-
-const TokenDataContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  height: 100%;
-  width: 100%;
 `
 
 const NoTokenDisplay = styled.div`
