+++ src/components/PositionTable/comonStyle.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components/macro'
+
+export const TokenDataContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  padding: 10px;
+  height: 100%;
+  width: 100%;
+`
